thong update code

+++ controller/FormController.js
@@ -3,15 +3,12 @@ const api = "http://localhost:3000/sv";
 function formController($scope, $http) {
   $scope.students = [];
   $scope.isLoading = false;
-  $scope.isSuccess = true;
+  $scope.isSuccess = false;
   $scope.message = "";
   $scope.index = -1;
-  $scope.sv = { gender: "1"};
+  $scope.sv = { gender: "1" };
 
 
-  $scope.onFormSubmit = function (event) {
-    event.preventDefault();
-  }
   $scope.isLoading = true;
   $http.get(api) // Gửi 1 request GET tới API
     .then(function (response) { // Phản hồi của API
@@ -24,6 +21,18 @@ function formController($scope, $http) {
       $scope.isLoading = false;
     });
 
+  function getapi() {
+    $http.get(api) // Gửi 1 request GET tới API
+      .then(function (response) { // Phản hồi của API
+        $scope.students = response.data;
+        console.log(response);
+        $scope.isLoading = false;
+      })
+      .catch(function (error) {
+        console.log(error);
+        $scope.isLoading = false;
+      });
+  }
 
   $scope.filltoForm = function (index) {
     Update = api + "/" + $scope.students[index].id;
@@ -32,8 +41,8 @@ function formController($scope, $http) {
     $scope.sv = angular.copy($scope.students[index]);
     console.log($scope.students[index].id);
   }
-
-  $scope.onInsert = function () {
+  $scope.onInsert = function (event) {
+    event.preventDefault();
     console.log($scope.sv);
     if ($scope.sv.fullname != null &&
       $scope.sv.name != null &&
@@ -48,7 +57,7 @@ function formController($scope, $http) {
             console.log(response);
             $scope.message = "Thêm mới thành công";
             $scope.isSuccess = true;
-            alert("Thêm Mới Thành Công");
+            getapi();
           })
       } else {
         alert("Mật Khẩu không trùng với Mật Khẩu Xác Nhận");
@@ -57,27 +66,46 @@ function formController($scope, $http) {
       alert("Vui lòng điền đầy đủ thông tin trên Form");
     }
   }
+  $scope.onSubmitForm = function (event) {
+    event.preventDefault();
+
+    // Gửi 1 request POST tới API
+    $http.post(api, $scope.sv)
+      .then(function (response) {
+        $scope.message = "Thêm mới thành công";
+        $scope.isSuccess = true;
+      });
+  }
 
   $scope.onUpdate = function () {
     $http.put(Update, $scope.sv)
       .then(function (response) {
         console.log(response.data);
-        alert("Sửa Thành Công");
+        $scope.message = "Sửa thành công";
+        $scope.isSuccess = true;
       })
   }
 
 
   $scope.onDelete = function () {
     console.log(Update);
-    $http.delete(Update, $scope.sv)
-      .then(function (response) {
-        console.log(response)
-        alert("Xóa Thành Công");
-      })
+      if ($scope.sv.name == "Admin") {
+        $scope.message="Bạn không thể xóa tài khoản Admin";
+        $scope.isSuccess = true;
+        return;
+      } else {
+        $http.delete(Update, $scope.sv)
+          .then(function (response) {
+            console.log(response)
+            $scope.message = "Xoá thành công";
+            $scope.isSuccess = true;
+          })
+    }
   }
 
   $scope.onClear = function () {
     $scope.index = -1;
     $scope.sv = { gender: "1" };
+    $scope.message = "";
   }
 }
